Moving everything related to Quesma v1 http proxy to dual_write_proxy.go
